Changed the way the command line arguments for sonar-runner are built.
Previously when the values where not defined default values where sent
to sonar-runner. For instance not setting a 'instance.hostUrl' would
use a default hostUrl of localhost:9000.

Now when a value is undefined it is not added as a parameter to
sonar-runner. This way the default configuration of sonar if self is
used as defined on the system where sonar-runner is called.

This is handy for when you don't want to commit passwords into your
git repository.

Fixes #22